[Toyota US] Fix spider
